+++ src/SFA.DAS.EmployerIncentives.Domain/Factories/IncentiveApplicationFactory.cs
@@ -12,7 +12,7 @@ namespace SFA.DAS.EmployerIncentives.Domain.Factories
             return IncentiveApplication.New(id, accountId, accountLegalEntityId);
         }
 
-        public Apprenticeship CreateNewApprenticeship(int apprenticeshipId, string firstName, string lastName, DateTime dateOfBirth, long uln, DateTime plannedStartDate, ApprenticeshipEmployerType apprenticeshipEmployerTypeOnApproval)
+        public Apprenticeship CreateNewApprenticeship(long apprenticeshipId, string firstName, string lastName, DateTime dateOfBirth, long uln, DateTime plannedStartDate, ApprenticeshipEmployerType apprenticeshipEmployerTypeOnApproval)
         {
             return new Apprenticeship(Guid.NewGuid(), apprenticeshipId, firstName, lastName, dateOfBirth, uln, plannedStartDate, apprenticeshipEmployerTypeOnApproval);
         }
